feat: add reusable page utilities
